+++ packages/amesos2/src/Amesos2_Factory.cpp
@@ -128,6 +128,14 @@ namespace Amesos2 {
         solverName == "pardisomkl") {
       return true;
     }
+#ifdef HAVE_MPI
+    if (solverName == "amesos2_css_mkl" ||
+        solverName == "css_mkl" ||
+        solverName == "amesos2_cssmkl"  ||
+        solverName == "cssmkl") {
+      return true;
+    }
+#endif
 #endif
 
 #ifdef HAVE_AMESOS2_LAPACK