+++ packages/kokkos-kernels/sparse/src/KokkosSparse_spmv_handle.hpp
@@ -36,8 +36,11 @@ enum SPMVAlgorithm {
                     /// is only used once.
   SPMV_NATIVE,      /// Use the best KokkosKernels implementation, even if a TPL
                     /// implementation is available.
-  SPMV_MERGE_PATH,  /// Use load-balancing merge path algorithm (for CrsMatrix
-                    /// only)
+  SPMV_MERGE_PATH,  /// Use algorithm optimized for matrices with
+                    /// imbalanced/irregular sparsity patterns (merge path or
+                    /// similar). May call a TPL. For CrsMatrix only.
+  SPMV_NATIVE_MERGE_PATH,  /// Use the KokkosKernels implementation of merge
+                           /// path. For CrsMatrix only.
   SPMV_BSR_V41,  /// Use experimental version 4.1 algorithm (for BsrMatrix only)
   SPMV_BSR_V42,  /// Use experimental version 4.2 algorithm (for BsrMatrix only)
   SPMV_BSR_TC    /// Use experimental tensor core algorithm (for BsrMatrix only)
@@ -59,6 +62,7 @@ inline const char* get_spmv_algorithm_name(SPMVAlgorithm a) {
     case SPMV_FAST_SETUP: return "SPMV_FAST_SETUP";
     case SPMV_NATIVE: return "SPMV_NATIVE";
     case SPMV_MERGE_PATH: return "SPMV_MERGE_PATH";
+    case SPMV_NATIVE_MERGE_PATH: return "SPMV_NATIVE_MERGE_PATH";
     case SPMV_BSR_V41: return "SPMV_BSR_V41";
     case SPMV_BSR_V42: return "SPMV_BSR_V42";
     case SPMV_BSR_TC: return "SPMV_BSR_TC";
@@ -73,10 +77,11 @@ inline const char* get_spmv_algorithm_name(SPMVAlgorithm a) {
 inline bool is_spmv_algorithm_native(SPMVAlgorithm a) {
   switch (a) {
     case SPMV_NATIVE:
-    case SPMV_MERGE_PATH:
+    case SPMV_NATIVE_MERGE_PATH:
     case SPMV_BSR_V41:
     case SPMV_BSR_V42:
     case SPMV_BSR_TC: return true;
+    // DEFAULT, FAST_SETUP and MERGE_PATH may call TPLs
     default: return false;
   }
 }
@@ -352,6 +357,7 @@ struct SPMVHandle
     } else {
       switch (get_algorithm()) {
         case SPMV_MERGE_PATH:
+        case SPMV_NATIVE_MERGE_PATH:
           throw std::invalid_argument(std::string("SPMVHandle: algorithm ") +
                                       get_spmv_algorithm_name(get_algorithm()) +
                                       " cannot be used if A is a BsrMatrix");